refactor: group all server selection related code into a module
