+++ lib/core/sdam/common.js
@@ -28,11 +28,32 @@ const ServerType = {
   Unknown: 'Unknown'
 };
 
+const TOPOLOGY_DEFAULTS = {
+  useUnifiedTopology: true,
+  localThresholdMS: 15,
+  serverSelectionTimeoutMS: 30000,
+  heartbeatFrequencyMS: 10000,
+  minHeartbeatFrequencyMS: 500
+};
+
+function drainTimerQueue(queue) {
+  queue.forEach(clearTimeout);
+  queue.clear();
+}
+
+function clearAndRemoveTimerFrom(timer, timers) {
+  clearTimeout(timer);
+  return timers.delete(timer);
+}
+
 module.exports = {
   STATE_CLOSING,
   STATE_CLOSED,
   STATE_CONNECTING,
   STATE_CONNECTED,
+  TOPOLOGY_DEFAULTS,
   TopologyType,
-  ServerType
+  ServerType,
+  drainTimerQueue,
+  clearAndRemoveTimerFrom
 };
